python_chals/requirements.txt
uvicorn
starlette
jinja2
python-multipart
slowapi
pyjwt
pycryptodome